Extract cleanup audit recorder
